+++ tests/_bootstrap.php
@@ -1,0 +1,2 @@
+<?php
+use erusev\parsedown;